Fix for indexes on custom object types that do not resolve to any AttributeType but implement Comparable. Key change is at IndexImpl.saveEntryIndex (lines 80-85). Even if attribute type is null we can still use an index. Test: QueryIndexTest.testResultsReturned_whenCustomAttributeIndexed Also, split up QueryBasicTest into multiple files.
