+++ hazelcast/src/main/java/com/hazelcast/query/impl/IndexImpl.java
@@ -18,6 +18,7 @@ package com.hazelcast.query.impl;
 
 import com.hazelcast.nio.serialization.Data;
 import com.hazelcast.query.QueryException;
+import com.hazelcast.query.impl.TypeConverters.TypeConverter;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -40,7 +41,7 @@ public class IndexImpl implements Index {
     private final String attribute;
     private final boolean ordered;
 
-    private volatile AttributeType attributeType;
+    private volatile TypeConverter converter;
 
     public IndexImpl(String attribute, boolean ordered) {
         this.attribute = attribute;
@@ -60,8 +61,8 @@ public class IndexImpl implements Index {
     public void clear() {
         recordValues.clear();
         indexStore.clear();
-        // Clear attribute type
-        attributeType = null;
+        // Clear converter
+        converter = null;
     }
 
     ConcurrentMap<Data, QueryableEntry> getRecordMap(Comparable indexValue) {
@@ -71,16 +72,18 @@ public class IndexImpl implements Index {
     @Override
     public void saveEntryIndex(QueryableEntry e) throws QueryException {
         /*
-         * At first, check if attribute type is not initialized, initialize it before saving an entry index
+         * At first, check if converter is not initialized, initialize it before saving an entry index
          * Because, if entity index is saved before,
-         * that thread can be blocked before executing attribute type setting code block,
-         * another thread can query over indexes without knowing attribute type and
+         * that thread can be blocked before executing converter setting code block,
+         * another thread can query over indexes without knowing the converter and
          * this causes to class cast exceptions.
          */
-        if (attributeType == null) {
+        if (converter == null) {
             // Initialize attribute type by using entry index
-            attributeType = e.getAttributeType(attribute);
+            AttributeType attributeType = e.getAttributeType(attribute);
+            converter = attributeType == null ? new IdentityConverter() : attributeType.getConverter();
         }
+
         Data key = e.getIndexKey();
         Comparable oldValue = recordValues.remove(key);
         Comparable newValue = e.getAttribute(attribute);
@@ -96,22 +99,20 @@ public class IndexImpl implements Index {
         } else {
             // update
             indexStore.updateIndex(oldValue, newValue, e);
-            //indexStore.removeIndex(oldValue, key);
-            //indexStore.newIndex(newValue, e);
         }
     }
 
     @Override
     public Set<QueryableEntry> getRecords(Comparable[] values) {
         if (values.length == 1) {
-            if (attributeType != null) {
+            if (converter != null) {
                 return indexStore.getRecords(convert(values[0]));
             } else {
                 return new SingleResultSet(null);
             }
         } else {
             MultiResultSet results = new MultiResultSet();
-            if (attributeType != null) {
+            if (converter != null) {
                 Set<Comparable> convertedValues = new HashSet<Comparable>(values.length);
                 for (Comparable value : values) {
                     convertedValues.add(convert(value));
@@ -124,7 +125,7 @@ public class IndexImpl implements Index {
 
     @Override
     public Set<QueryableEntry> getRecords(Comparable value) {
-        if (attributeType != null) {
+        if (converter != null) {
             return indexStore.getRecords(convert(value));
         } else {
             return new SingleResultSet(null);
@@ -134,7 +135,7 @@ public class IndexImpl implements Index {
     @Override
     public Set<QueryableEntry> getSubRecordsBetween(Comparable from, Comparable to) {
         MultiResultSet results = new MultiResultSet();
-        if (attributeType != null) {
+        if (converter != null) {
             indexStore.getSubRecordsBetween(results, convert(from), convert(to));
         }
         return results;
@@ -143,17 +144,14 @@ public class IndexImpl implements Index {
     @Override
     public Set<QueryableEntry> getSubRecords(ComparisonType comparisonType, Comparable searchedValue) {
         MultiResultSet results = new MultiResultSet();
-        if (attributeType != null) {
+        if (converter != null) {
             indexStore.getSubRecords(results, comparisonType, convert(searchedValue));
         }
         return results;
     }
 
     private Comparable convert(Comparable value) {
-        if (attributeType == null) {
-            return value;
-        }
-        return attributeType.getConverter().convert(value);
+        return converter.convert(value);
     }
 
     public ConcurrentMap<Data, Comparable> getRecordValues() {
