+++ hazelcast/src/test/java/com/hazelcast/map/query/CustomAttribute.java
@@ -1,0 +1,53 @@
+/*
+ * Copyright (c) 2008-2013, Hazelcast, Inc. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.hazelcast.map.query;
+
+import com.hazelcast.util.HashUtil;
+
+import java.io.Serializable;
+
+public class CustomAttribute implements Serializable, Comparable<CustomAttribute> {
+
+    private int age;
+    private long height;
+
+    public CustomAttribute(int age, long height) {
+        this.age = age;
+        this.height = height;
+    }
+
+    public int compareTo(CustomAttribute o) {
+        return age - o.age;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+
+        CustomAttribute that = (CustomAttribute) o;
+
+        if (age != that.age) return false;
+        if (height != that.height) return false;
+
+        return true;
+    }
+
+    @Override
+    public int hashCode() {
+        return HashUtil.hashCode(age, height);
+    }
+}
